requirements.txt
#in ubuntu install libmysqlclient-dev
Django==1.3
MySQL-python==1.2.4
django-debug-toolbar==0.9.4
ipython==0.13.2
unipath==1.0
django-taggit==0.10a1
South==0.8.1
django-braces
